requirements.txt
# Core dependencies
pdf2image==1.17.0
Pillow==11.0.0
reportlab==4.0.4
docx2pdf==0.1.8
python-dateutil==2.9.0.post0

# Optional dependencies (commented out as not needed for offline version)
# boto3==1.35.86
# Flask==3.1.0
# Flask-Cors==5.0.0
# Flask-SocketIO==5.5.0
# gevent==24.11.1
# python-socketio==5.12.0

# System dependencies (install via package manager)
# poppler-utils  # For pdf2image
# libreoffice    # For docx2pdf
